refactor: group AI spacing shape modules
